refactor: change apis structure
