+++ packages/backend/src/index.ts
@@ -2,10 +2,12 @@ import fastifyCookie from "@fastify/cookie";
 import fastifySession from "@fastify/session";
 import Fastify from "fastify";
 import config from "../config.json";
+
 import mongoPlugin from "./plugins/mongo";
-import apiRouter from "./routes/api";
-import indexRouter from "./routes/index";
+
+import { coursesRouter, usersRouter } from "./routes/apis";
 import authRouter from "./routes/auth";
+import indexRouter from "./routes/index";
 
 const fastify = Fastify({ logger: !config.prod });
 
@@ -19,9 +21,10 @@ fastify.register(fastifySession, {
 	},
 });
 
-fastify.register(indexRouter);
-fastify.register(apiRouter);
 fastify.register(authRouter);
+fastify.register(coursesRouter);
+fastify.register(indexRouter);
+fastify.register(usersRouter);
 
 fastify.listen({ host: "0.0.0.0", port: config.port }, err => {
 	if (err) {
