+++ packages/backend/src/routes/apis.ts
@@ -1,0 +1,4 @@
+import usersRouter from "./apis/users";
+import coursesRouter from "./apis/courses";
+
+export { usersRouter, coursesRouter };
